requirements.txt
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
PyYAML==6.0.1
rumps==0.4.0
